Add non-interactive mode
A new command line option --non-interactive may be used to
prevent prompts, using defaults, and generating a .coafile
that is safe to use in non-interactive mode, as it disables
all bears needing values for non-optional settings.

The introduction of a command line parser requires that
the parser is given to coalib.settings functions, otherwise
it would instantiate the default coala command line parser
which has many arguments that are either not appropropriate
or not implemented in coala-quickstart.

Closes https://github.com/coala/coala-quickstart/issues/39