tests/Bears.py: Update npm package versions
Updates the version of npm dependencies in the string used
to generate package.json while executing some of the tests.
The version needed to be changed for the tests to pass because
of the upstream upgrade in versions of Bear dependencies in
coala-bears.

Fixes https://github.com/coala/coala-quickstart/issues/194